% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnets.R
\name{best_subnets}
\alias{best_subnets}
\title{Find the optimal sub-networks}
\usage{
best_subnets(func, net.matrix, max.size = 10, exhaust.size = 5,
  max.top = 10000)
}
\description{
Find the optimal sub-networks
}

