% Generated by roxygen2 (4.0.1): do not edit by hand
\name{build_sim_net}
\alias{build_sim_net}
\title{Build the gene similarity network from annoations}
\usage{
build_sim_net(core.sets, kappa.threshold = 0.9)
}
\arguments{
\item{core.sets}{Logical matrix indicated the core genes associated with specific functions or pathways.
The rows correspond to genes, while the columns represent the functions or pathways.}

\item{kappa.threshold}{The threshold to establish a link between two genes if their kappa score is larger
than the threshold.}
}
\value{
This function returns a sparse adjacent matrix of the gene similarity network.
}
\description{
Build the gene similarity network from the core sets matrix by using Cohen's kappa statistic.
}
\details{
This function build the gene similarity network from the gene annotations, which is given
by a core sets matrix. The Cohen's kappa statistic is used to compute the similarity between
two genes.
}

