% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnets.R
\name{extend_subnets}
\alias{extend_subnets}
\title{Extend sub-networks by overlapping two sub-networks}
\usage{
extend_subnets(subnet1, subnet2, size = 0)
}
\description{
Extend sub-networks by overlapping two sub-networks
}

