% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{get_adjusted_deg_diff}
\alias{get_adjusted_deg_diff}
\title{Calculate adjusted degree differences for given network}
\usage{
get_adjusted_deg_diff(net, log.expr.val, p = 0.5)
}
\description{
Calculate adjusted degree differences for given network
}
