% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{get_diff_ratio_net}
\alias{get_diff_ratio_net}
\title{Construct differential expression ratio network}
\usage{
get_diff_ratio_net(ref.ratio.dist, expr.val, log.expr = FALSE)
}
\description{
Construct differential expression ratio network
}
