% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{get_ratio_distribution}
\alias{get_ratio_distribution}
\title{Calculate expression ratio distribution}
\usage{
get_ratio_distribution(ref.expr.matrix, p.edge = 0.1, log.expr = FALSE)
}
\description{
Calculate expression ratio distribution
}
