% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{get_ratio_distribution2}
\alias{get_ratio_distribution2}
\title{Calculate expression ratio distribution}
\usage{
get_ratio_distribution2(ref.expr.matrix, p.edge = 0.1, p.trim = 0.3,
  log.expr = FALSE, use.parallel = FALSE)
}
\description{
Calculate expression ratio distribution
}
