% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_normalization.R
\name{get_ratio_variance}
\alias{get_ratio_variance}
\title{Calculate expression ratio variances}
\usage{
get_ratio_variance(expr.matrix, log.expr = FALSE)
}
\arguments{
\item{expr.matrix}{The expression matrix. Each row represents a gene and each column represents a sample.}

\item{log.expr}{Logical variable indicating whether the input expression matrix is in logarithmic scale.}
}
\value{
This function will return a numeric matrix with element [i,j] represents the variance of 
expressioin ratios for gene pairs (i, j).
}
\description{
Calculate the variances of expression ratios for each pair of genes.
}
