% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_normalization.R
\name{get_stable_genes}
\alias{get_stable_genes}
\title{Identify the stable genes}
\usage{
get_stable_genes(
  expr.matrix,
  max.stable,
  p.edge = 0.25,
  log.expr = FALSE,
  use.parallel = FALSE
)
}
\arguments{
\item{expr.matrix}{The expression matrix. Each row represents a gene and each column represents a sample.}

\item{max.stable}{The maximal number of genes that are selected as the stable genes.}

\item{p.edge}{The percentage of gene pairs that are selected into the uniform ratio graph.}

\item{log.expr}{Logical variable indicating whether the input expression matrix is in logarithmic scale.}

\item{use.parallel}{Logical variable indicating to use the BiocParallel package to accelerate computation.}
}
\value{
This function will return an integer vector containing the indexes of stable genes.
}
\description{
Identify the stable genes by using URG (uniform ratio graph) method.
}
