% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kappa_score}
\alias{kappa_score}
\title{Cohen's kappa score}
\usage{
kappa_score(x1, x2)
}
\arguments{
\item{x1}{The first logical vector}

\item{x2}{The second logical vector}
}
\value{
The Cohen's kappa score
}
\description{
Calculate Cohen's kappa score for two vectors.
}
\details{
This function calculate Cohen's kappa score for two logical vectors.
}
