\name{neeat}
\alias{neeat}
\title{Network enhanced enrichment analysis tool}
\usage{
neeat(w, r)
}
\description{
Functional enrichment anslysis of gene sets by integrating
network information
}
\details{
This is a novel functional enrichment analysis tool based
on the gene set and network information.
}

