\name{neeat_g}
\alias{neeat_g}
\title{Network enhanced enrichment analysis tool}
\usage{
neeat_g(gene.set, core.set, net, rho = 0.5, n.perm = 10000)
}
\description{
Functional enrichment anslysis of gene sets by integrating
network information
}
\details{
This is a novel functional enrichment analysis tool based
on the gene set and network information.
}

