% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{netDEG}
\alias{netDEG}
\title{netDEG: Differentially expressed gene identification method}
\usage{
netDEG(ref.expr.matrix, expr.matrix, p.edge = 0.1, log.expr = FALSE,
  summarize = c("gene", "sample"), use.parallel = FALSE)
}
\description{
netDEG: Differentially expressed gene identification method
}
