% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{netDEG_pvalue}
\alias{netDEG_pvalue}
\title{Calculate netDEG statistics and p-values}
\usage{
netDEG_pvalue(ref.ratio.dist, expr.val, log.expr = FALSE)
}
\description{
Calculate netDEG statistics and p-values
}
