% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihyper.R
\name{pmultihyper}
\alias{pmultihyper}
\title{The Multivariate Hypergeometric Distribution}
\usage{
pmultihyper(x, k, m, w)
}
\arguments{
\item{x}{The quantile of weighted sum.}

\item{k}{The total number of balls drawn from the urn.}

\item{m}{Integer non-negative vector of length N, containing the number of balls of each color in the urn.
N is the number of colors.}

\item{w}{Numeric non-negative vector of length N, specifying the weight of balls of each color.}
}
\value{
This function will return the probablity of \eqn{P(X \le x)}.
}
\description{
The distribution function for the weighted sums of multivariate hypergeometric distribution
}
\details{
This function gives the distribution function for the weighted sums of multivariate hypergeometric
distribution by recursively calling the hypergeometric distribution density function
\code{\link{dhyper}}.
}
\seealso{
\code{\link{dhyper}}
}
