% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_simulation.R
\name{simulate_dropout}
\alias{simulate_dropout}
\title{Simulate dropout expression data}
\usage{
simulate_dropout(counts, dropout.rate = 0, dropout.rate.sd = 0.1)
}
\description{
Simulate dropout expression data
}
\references{
Peter V. Kharchenko, Lev Silberstein, and David T. Scadden.
Bayesian approach to single-cell differential expression analysis.
Nature Methods, 11(7):740–742, 2014.
}
