% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_simulation.R
\name{simulate_sample_groups}
\alias{simulate_sample_groups}
\title{Simulate sample groups from given samples with labels}
\usage{
simulate_sample_groups(labels, groups, sizes, replace = FALSE)
}
\description{
Simulate sample groups from given samples with labels
}
