% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{submatrix}
\alias{submatrix}
\title{Extract a submatrix from a matrix}
\usage{
submatrix(m, rows, cols)
}
\arguments{
\item{m}{The matrix}

\item{rows}{The integer vectors of row index(es)}

\item{cols}{The integer vectors of column index(es)}
}
\value{
This function will return the specified submatrix as a matrix of corresponding type.
}
\description{
Extract a specified submatrix from a sparse matrix rapidly
}
\details{
This function implements faster submatrix extraction algorithm for the 
\code{\link[=CsparseMatrix-class]{CsparseMatrix}} class in the package \pkg{Matrix}.
}

