% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toMatrix}
\alias{toMatrix}
\title{Turn a two-columns mapping table to matrix}
\usage{
toMatrix(table, rows = unique(table[, 1]), cols = unique(table[, 2]))
}
\arguments{
\item{table}{The two-columns table as matrix or data frame. The other columns will not be used, if available.}

\item{rows}{The row names for the mapping matrix.}

\item{cols}{The column names for the mapping matrix.}
}
\value{
This function returns a sparse logical matrix.
}
\description{
Generate the sparse mapping matrix from a two-columns table.
}
\details{
This function generates the sparse logical matrix from the two-columns table returned by \code{\link{toTable}}
in Bioconductor package \code{\link{AnnotationDbi}}.
}
\examples{
\dontrun{
source("http://bioconductor.org/biocLite.R")
biocLite("org.Hs.eg.db")
library(org.Hs.eg.db)
x <- toMatrix(toTable(org.Hs.egGO2ALLEGS))
}
}
\seealso{
\code{\link{toTable}}
}

