% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_normalization.R
\name{URG_normalize}
\alias{URG_normalize}
\title{Normalize using given factors}
\usage{
URG_normalize(expr.matrix, factor, log.expr = FALSE)
}
\arguments{
\item{expr.matrix}{The expression matrix. Each row represents a gene and each column represents a sample.}

\item{factor}{The numeric vector of normalization factors.}

\item{log.expr}{Logical variable indicating whether the input expression matrix is in logarithmic scale.}
}
\value{
This function will return a numeric matrix with the same dimension of \code{expr.matrix}.
}
\description{
Normalize the expression matrix by using the given factor for each sample.
}
\seealso{
\code{\link{URG_getFactor}}
}
