% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{get_adjusted_deg_diff}
\alias{get_adjusted_deg_diff}
\title{Calculate adjusted degree differences for given network}
\usage{
get_adjusted_deg_diff(net, log.expr.val, p = 0.5)
}
\arguments{
\item{net}{The binary adjacent matrix of differential expression ratio network.}

\item{log.expr.val}{Numeric vector containing the logarithmic scale gene expression values.}

\item{p}{The parameter for calculating the adjusted degree differences.}
}
\value{
This function will return a list with the following components:
  \item{diff}{A numeric vector containing the adjusted degree differences of all genes.}
  \item{degree}{A list containing the raw degree differences and sums of all genes.}
}
\description{
Calculate the adjusted degree differences for all genes in the given network.
}
