% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neeat}
\alias{neeat}
\title{Network enhanced enrichment analysis tool}
\usage{
neeat(core.sets, gene.sets = NULL, net = NULL, subnet = NULL,
  depths = NULL, method = "gene", rho = 0.001, max.depth = 2,
  n.perm = 0, use.multinom = FALSE, z.threshold = 0, verbose = FALSE,
  adjust.p = "BH", n.cpu = 1, batch.size = 5000)
}
\arguments{
\item{core.sets}{Logical matrix indicated the core genes associated with specific functions or pathways.
The rows correspond to genes, while the columns represent the functions or pathways.}

\item{gene.sets}{Logical vector or matrix indicated the gene set(s) for evaluating.
The rows correspond to genes, while the columns represent the gene sets.}

\item{net}{The adjacent matrix of network.}

\item{subnet}{The adjacent matrix of sub-network for evaluating in the NEEAT-subnet model.}

\item{depths}{The node depths defined in the NEEAT model, which can be pre-computed by calling
\code{\link{neeat_depths}} in advance in order to reduce the computation time. If not provided,
it will be computed on-demand.}

\item{method}{A string indicated the NEEAT model, including "gene", "net" and "subnet".
Use "hyper" for traditional hypergeometric test, in which the network information is ignored.}

\item{rho}{The weight parameter for depths.}

\item{max.depth}{Integer for the maximum depth considered in the NEEAT models.}

\item{n.perm}{The number of permutations for calculating p-values.}

\item{use.multinom}{Logical variable indicated whether use \code{\link{rmultinom}} to
approximate \code{\link{rmultihyper}}.}

\item{z.threshold}{The threshold for filtering out small Z-scores. The p-values are calculated only
for the Z-scores \code{>= z.threshold}, otherwise p-values are set as 1.}

\item{adjust.p}{The method for adjusting p-values for multiple testing. Use "none" for bypassing.
See \code{\link{p.adjust}} for available methods.}

\item{n.cpu}{The number of CPUs/cores used in the parallel computation.}

\item{batch.size}{The desired size of batches in the parallel computation.}
}
\value{
This function will return a 3-dimensional array of dimensions \code{c(5, dim(gene.sets)[2], dim(core.sets)[2])},
and each column \code{[,i,j]} containing the following components for the correponding gene set \code{gene.sets[,i]}
and core set \code{core.sets[,j]}:
  \item{\code{z.score}}{The Z-score for the gene set}
  \item{\code{p.value}}{The statistic significance for the gene set under specified NEEAT model}
  \item{\code{raw.score}}{The raw score for the gene set under specified NEEAT model}
  \item{\code{avg.score}}{The average score for random permutations of the gene set}
  \item{\code{var.score}}{The variance of scores for random permutations of the gene set}
}
\description{
Functional enrichment anslysis of gene sets by integrating network information
}
\details{
This is a novel functional enrichment analysis tool based on the gene set and network
information.
}
\seealso{
\code{\link{get_core_sets}}, \code{\link{neeat_depths}}
}

