% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{netDEG_pvalue}
\alias{netDEG_pvalue}
\title{Calculate netDEG p-values}
\usage{
netDEG_pvalue(ref.ratio.dist, expr.val, log.expr = FALSE)
}
\arguments{
\item{ref.ratio.dist}{The expression ratio distribution profile returned by \code{get_ratio_distribution} or \code{get_ratio_distribution2}.}

\item{expr.val}{Numeric vector of gene expression values in the sample.}

\item{log.expr}{Logical variable indicating whether the input expression vector is in logarithmic scale.}
}
\value{
This function will return a list with the following components:
  \item{up}{A numeric vector containing the p-values of up-regulation test.}
  \item{down}{A numeric vector containing the p-values of down-regulation test.}
  \item{twoside}{A numeric vector containing the p-values of twoside test.}
}
\description{
Perform the single or two side tests and calculate the p-values.
}
