% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_rationet.R
\name{p_combine}
\alias{p_combine}
\title{Calculate combined p-value by Fisher's method}
\usage{
p_combine(p)
}
\arguments{
\item{p}{the numeric vector containing the p-values need to combine.}
}
\value{
This function will return a list with the following components:
  \item{chisq}{The chi-squared statistic.}
  \item{df}{The degrees of freedom of chi-squared distribution.}
  \item{p}{The p-value of chi-squared statistic.}
}
\description{
The Fisher's method is used to combine the results from several independent tests.
}
