% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihyper.R
\name{rmultihyper}
\alias{rmultihyper}
\title{The Multivariate Hypergeometric Distribution}
\usage{
rmultihyper(n, k, m)
}
\arguments{
\item{n}{The number of observations.}

\item{k}{The total number of balls drawn from the urn.}

\item{m}{The integer vector containing the number of balls of each color in the urn.
Length of vector is the number of colors.}
}
\value{
This function will return a matrix of \code{length(m)} rows and \code{n} columns,
and each column contains the number of balls of each color drawn from the urn.
}
\description{
Generate random variables for the multivariate hypergeometric distribution
}
\details{
This function generates random variables for the multivariate hypergeometric
distribution by iteratively calling hypergeometric random variable generator
\code{\link{rhyper}}.
}
\seealso{
\code{\link{rhyper}}
}
